add integration test for sqlserver
Refactored the pom so that deployment modules are only run
for the 'deploy' profile. This is so that they do not run during
the 'verify' maven phase.

Added an integration test that will start up sqlserver in a
docker container on an ephemeral port before the test,
then stop it after the test. The port information is written to a
file, then read at the start of the integration test.