add mysql integration test
Added an integration test for MySQL plugin that starts up a
Docker image at the start and tears it down at the end.

Also did some refactoring in the pom to remove the maven plugin
that writes pom properties to a file, in favor of more transparently
setting system properties in the failsafe plugin itself.