CDAP-16737 fix stop logic
Fixed stop logic for the event readers. Event consumers need to
throw a special StopConnectorException to tell Debezium to stop
immediately instead of continuing. Fixed the consumers to do this
if they are interrupted while writing to the event queue, which
happens when the reader is being stopped.

Also added integration tests for stopping logic and
included fixes to mysql integration test